refactor(capreg): split 898-line store.go by concern (core→51)
store.go bundled the Store contract, all domain types, the in-memory
store, and the canonicalization/hashing layer. Split into cohesive
sibling files within package capreg — pure relocation, no call-site or
behavior changes:

- store.go (51): Store interface, identity types, NewStore factory
- capreg_types.go: domain value types + lookup interfaces
- capreg_memory.go: MemoryStore impl + feedback-summary helpers
- capreg_canonical.go: canonicalization, fingerprinting, hashing

41 funcs in, 41 funcs out (lossless). Verified by go build, go vet,
the capreg tests (fingerprint/canonical golden), scoped golangci-lint,
goimports.
